Do not assume a systemd daemon is running after install
If the package is installed as part of a cdebootstrap environment (a fairly common use case), then the systemd daemon is not running at all. This commit removes the 2 commands that depend on the daemon existing, whilst still ensuring the service will be enabled.

Debian itself has better ways to manage systemd services, so this approach is the least worse?